bookItinerary and cancelItinerary operations added with their logic (except fault handling)
bookItinerary and cancelItinerary operations have been added. They are
missing the part to handle the faults.
The service builds and deploys correctly but there is an error when
refreshing the server on client side. It is due to the
creditCardInfoType, which is imported two times. This is because we
have some problem in the XML schema